// src/app/Models/Cliente.ts
export class Cliente {
  nombre1: string;
  nombre2: string;
  apellido1: string;
  apellido2: string;
  cedula: string;
  fechanacimiento: Date;
  edad: number;
  peso: number;
  imc: number;
  correo: string;
  password: string;
  ciudad: string;
  codigopostal: number;
  indicaciones: string;
  constructor() {
  }
}